Fixed the code to successfully compare acceleration values
Before these fixes, acceleration values were not able to be
compared, as they needed to be unpacked from the sensors_vec_t.v
array. x = v[0] and y = v[1]. I also changed the comparators to
use the int 2 instead of looking for an undefined type "2G".

Let me know if this works better, as the other file had a lot of
errors!